pyproject.toml
[project]
name = "b-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.2"
dependencies = [
    "aiohttp>=3.11.18",
    "asyncclick>=8.1.8",
    "asyncio>=3.4.3",
    "click>=8.1.8",
    "fastapi>=0.115.12",
    "hf-xet>=1.1.0",
    "httpx>=0.28.0",
    "httpx-sse>=0.4.0",
    "huggingface-hub>=0.30.2",
    "jwcrypto>=1.5.6",
    "langchain>=0.3.25",
    "langchain-community>=0.3.23",
    "langchain-core>=0.3.58",
    "langchain-ollama>=0.3.2",
    "matplotlib>=3.10.3",
    "mysql-connector-python>=9.3.0",
    "nuitka>=2.7.2",
    "numpy>=2.2.5",
    "openai>=1.78.1",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pipeline>=0.1.0",
    "pydantic>=2.11.4",
    "pyhocon>=0.3.61",
    "pyinstaller>=6.13.0",
    "pyjwt>=2.10.1",
    "pymupdf>=1.25.5",
    "pysnmplib>=5.0.24",
    "python-dotenv>=1.1.0",
    "ray[all]>=2.47.1",
    "requests>=2.32.3",
    "ruff>=0.13.3",
    "snmpclitools>=0.7.2",
    "starlette>=0.46.2",
    "streamlit>=1.45.1",
    "timm>=1.0.15",
    "torch>=2.7.0",
    "torchvision>=0.20.1",
    "uvicorn>=0.34.2",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
